refactor(layout): переносит логику компонента в отдельный хук
